move things around and rewrite to typescript
