+++ package.json
@@ -1,9 +1,12 @@
 {
   "name": "@allthings/react-intl-extract-messages",
   "version": "0.2.0",
-  "bin": "cli.js",
-  "main": "index.js",
+  "bin": "dist/cli.js",
+  "main": "dist/index.js",
   "license": "MIT",
+  "files": [
+    "dist"
+  ],
   "engines": {
     "node": ">=8"
   },
@@ -18,17 +21,34 @@
     "tmp": "0.0.33"
   },
   "devDependencies": {
+    "@types/babel__core": "7.0.2",
+    "@types/babel__generator": "7.0.1",
+    "@types/babel__template": "7.0.1",
+    "@types/babel__traverse": "7.0.1",
+    "@types/glob": "7.1.1",
+    "@types/minimist": "1.2.0",
+    "@types/node": "10.12.9",
+    "@types/tmp": "0.0.33",
     "jest": "23.6.0",
-    "prettier": "1.15.2"
+    "prettier": "1.15.2",
+    "rimraf": "2.6.2",
+    "ts-jest": "23.10.4",
+    "typescript": "3.1.6"
   },
   "scripts": {
     "postversion": "git push origin master --follow-tags && npm publish",
     "pretest": "cd test/fixture-project && yarn --frozen-lockfile",
-    "test": "jest --coverage"
+    "test": "jest --coverage",
+    "prebuild": "rimraf dist",
+    "build": "tsc -p tsconfig.json"
   },
   "prettier": {
     "semi": false,
     "singleQuote": true,
     "trailingComma": "es5"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node"
   }
 }
