+++ src/index.ts
@@ -1,13 +1,26 @@
-const childProcess = require('child_process')
-const fs = require('fs')
-const path = require('path')
-const glob = require('glob')
-const tmp = require('tmp')
-const babel = require('@babel/core')
+import childProcess from 'child_process'
+import fs from 'fs'
+import path from 'path'
+
+import glob from 'glob'
+import tmp from 'tmp'
+import { transformFileSync, BabelFileMetadata } from '@babel/core'
 
 const projectDir = path.dirname(process.argv[1])
 
-function main(tsconfigFile) {
+interface IntlMessage {
+  id: string
+  defaultMessage: string
+  description?: string
+}
+
+interface BabelMetadataReactIntl extends BabelFileMetadata {
+  'react-intl': {
+    messages: IntlMessage[]
+  }
+}
+
+export default function main(tsconfigFile: string) {
   const tsconfigFilePath = path.isAbsolute(tsconfigFile)
     ? tsconfigFile
     : path.join(projectDir, tsconfigFile)
@@ -51,16 +64,23 @@ function main(tsconfigFile) {
       require('babel-plugin-react-intl'),
     ],
   }
-  const messages = files.reduce((allMessages, fileName) => {
-    const result = babel.transformFileSync(fileName, babelCompileOptions)
-    const { messages } = result.metadata['react-intl']
+  const messages: IntlMessage[] = files.reduce(
+    (allMessages: IntlMessage[], fileName) => {
+      const result = transformFileSync(fileName, babelCompileOptions)
 
-    return [...allMessages, ...messages]
-  }, [])
+      if (result && result.metadata) {
+        const metadata = result.metadata as BabelMetadataReactIntl
+        const { messages } = metadata['react-intl']
+
+        return [...allMessages, ...messages]
+      }
+
+      return allMessages
+    },
+    []
+  )
 
   tmpDir.removeCallback()
 
   return messages
 }
-
-module.exports = main
